+++ src/SparseDiffTools.jl
@@ -8,5 +8,7 @@ include("contraction_coloring.jl")
 include("greedy_d1_coloring.jl")
 include("matrix2graph.jl")
 
+include("program_sparsity/program_sparsity.jl")
+
 
 end # module
